web/requirements.txt
beautifulsoup4==4.4.1
click==6.6
coverage==4.0.3
Django==1.9.5
django-braces==1.8.1
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-filter==0.13.0
django-jenkins==0.18.1
django-rest-swagger==0.3.6
django-shotgun==0.2
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
drf-extensions==0.2.8
drf-hal-json==0.9.1
drf-nested-fields==0.9.3
cryptography==1.4
first==2.0.1
Markdown==2.6.6
netifaces==0.10.4
psycopg2==2.6.1
pyslack-real==0.6.0
python-dateutil==2.5.3
PyYAML==3.11
requests==2.9.1
six==1.10.0
sphinx-me==0.3
sqlparse==0.1.19
tqdm==4.4.3
urllib3==1.15.1
